+++ apps/chat/src/components/Common/FilesSelector/SelectedFile.tsx
